requirements.txt
amqp==5.1.1
anyio==3.6.1
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.0.7
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
black==22.6.0
cachetools==4.2.2

certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.7

django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-timezone-field==5.0
executing==0.9.1
flake8==5.0.4
gunicorn==20.1.0
h11
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.12.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
python-telegram-bot-django-persistence==0.1.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
